requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio==7.2.14
pandas>=1.5.0
boto3>=1.26.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.28.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-slack== 9.1.0